Editing the Concentration Camp Note and SCP-914 Test Log
I added a number to the note from the concentration camp (859305, this is the prisoner number that can be seen on his texture). Since there are 6 digits there, and the code for the door is only 4, then two random digits out of 6 should either be splattered with blood or crossed out using the game code (for example, a random code appears on the SCP-372 document, so I think making the random two digits covered by overlay text or blood is not a problem). I calculated that there will be 15 possible combinations of the code, so it is unlikely that players will be able to remember them all and will have to touch SCP-1123 to find out the code.

Also, I slightly modified one of SCP-914's test logs, adding an experiment to upgrade the radio, so players know that the radio needs to be upgraded to very fine.